refactor: extract scale props resolver
