+++ packages/core/src/view/scaleResolutionConstants.js
@@ -1,0 +1,5 @@
+export const QUANTITATIVE = "quantitative";
+export const ORDINAL = "ordinal";
+export const NOMINAL = "nominal";
+export const LOCUS = "locus";
+export const INDEX = "index";
